requirements.txt
streamlit==1.45.0
openai==1.77.0
python-dotenv==1.1.0
httpx==0.28.1
pandas==2.2.3
plotly==5.24.1
altair==5.5.0
faker==37.1.0
numpy==2.2.5
